requirements.txt
bidict==0.23.1
blinker==1.8.2
click==8.1.7
dnspython==2.6.1
eventlet==0.36.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SocketIO==5.3.7
greenlet==3.1.0
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
psutil==6.0.0
py-cpuinfo==9.0.0
python-engineio==4.9.1
python-socketio==5.11.4
setuptools==75.1.0
simple-websocket==1.0.0
Werkzeug==3.0.4
wsproto==1.2.0
